Self-play: ship the board as uint8 (4x less H2D), ~1.9x total over baseline
The board encoder writes a pure binary {0,1} tensor, so the netbatch collect() now emits it as
uint8 (4x smaller) and the GPU casts to float after the cheaper transfer; lines/glob keep their
fractional features in f32. The board was ~18MB of the ~19MB per-round host->device transfer,
which the profile under-counted — cutting it was a bigger lever than expected.

kd-features: encode_into is refactored into a generic encode_board<T>(.., one) + encode_aux
(lines/glob, always f32) and delegates to them, so f32 (training) and u8 (transfer) come from
ONE source. encode_into stays byte-identical (parity test unaffected); the net sees an identical
f32 {0,1} board, so forward outputs are bit-identical — this is transport-only.

Measured (RTX 4070S, 128 sims, random net, batch 1024), leaf-evals/s:
  single-pool:          95,070 -> 149,433  (1.57x from int8 alone)
  overlap (worker):    134,037 -> 180,367  (1.35x from int8; 1.90x total vs baseline)
Round is still GPU-path-bound (collect hidden); forward is now the largest remaining phase, so
CUDA graphs / torch.compile is the next lever.

+++ agent/kdagent/selfplay.py
@@ -170,7 +170,7 @@ def run_netbatch(args):
             total_leaves += b
             rounds += 1
             sync(); ta = time.perf_counter()
-            board = torch.from_numpy(batch["board"]).to(device, non_blocking=True)
+            board = torch.from_numpy(batch["board"]).to(device, non_blocking=True).float()
             lines = torch.from_numpy(batch["lines"]).to(device, non_blocking=True)
             glob = torch.from_numpy(batch["glob"]).to(device, non_blocking=True)
             if prof:
@@ -215,7 +215,7 @@ def _gpu_forward(net, batch, device, use_amp):
     by moving them to CPU, which is what lets a CPU `collect()` run concurrently with this."""
     import torch
 
-    board = torch.from_numpy(batch["board"]).to(device, non_blocking=True)
+    board = torch.from_numpy(batch["board"]).to(device, non_blocking=True).float()
     lines = torch.from_numpy(batch["lines"]).to(device, non_blocking=True)
     glob = torch.from_numpy(batch["glob"]).to(device, non_blocking=True)
     with torch.no_grad(), torch.autocast("cuda", dtype=torch.bfloat16, enabled=use_amp):
